Use decimals in stock amounts

+++ database/migrations/2020_02_12_100000_create_stock_mutations_table.php
@@ -18,7 +18,7 @@ class CreateStockMutationsTable extends Migration
             $table->morphs('stockable');
             $table->string('reference_type')->nullable();
             $table->unsignedBigInteger('reference_id')->nullable();
-            $table->integer('amount');
+            $table->decimal('amount');
             $table->text('description')->nullable();
             $table->timestamps();
